Created new DTO to return the 'createChapter' method and refactor dtos package.
