+++ manga_be/src/main/java/com/reader/manga/dto/chapter/GetChapterDTO.java
@@ -1,0 +1,7 @@
+package com.reader.manga.dto.chapter;
+
+public record GetChapterDTO(
+        String title,
+        String description,
+        Integer numberPages
+) {}
